Solving problems from memmove

+++ libft/ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: jmartos- <jmartos-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/01 12:49:31 by jmartos-          #+#    #+#             */
-/*   Updated: 2023/05/02 11:23:26 by jmartos-         ###   ########.fr       */
+/*   Updated: 2023/05/09 15:29:45 by jmartos-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 void	ft_bzero(void *s, size_t n)
 {
 	unsigned char	*str;
-	int				i;
+	size_t			i;
 
 	str = (unsigned char *)s;
 	i = 0;
